Added '-d', fixed file name chech, fixed typo

+++ hashname.py
@@ -20,22 +20,29 @@ def check_args():
         '-f',
         '--force',
         action='store_true',
-        help='Process the file even if it looks like it has already been processed '
+        help='Process the file even if it looks like it has already been processed'
     )
     
     parser.add_argument(
         '-F',
         '--force-rename',
         action='store_true',
-        help='Rename file even there is file with the same hash'
+        help='Rename file even there is another file with the same result name'
     )
-    
+
     parser.add_argument(
         '-a',
         '--algorithm',
         default='sha256',
         dest='algorithm',
-        help='Hash algorithm: "md5", "sha256" (default), "sha512"'
+        help='Set specific hash algorithm: "md5", "sha256" (default), "sha512"'
+    )
+
+    parser.add_argument(
+        '-d',
+        '--dry-run',
+        action='store_true',
+        help='Do not actually rename files'
     )
 
     parser.add_argument(
@@ -78,7 +85,8 @@ def get_algorithm(args):
 def is_already_hash(file_name: str, digest_size: int):
     file_name = os.path.splitext(file_name)[0]
 
-    return bool(re.match('^[a-f0-9]{64}$', file_name))
+    res = re.search('^[a-f0-9]{' + str(digest_size) + '}$', file_name)
+    return res is not None
 
 
 def process(args, algo):
@@ -86,11 +94,13 @@ def process(args, algo):
         hasher = algo()
 
         try:
+            # Is it a directory?
             if os.path.isdir(input_file):
                 if args.verbose:
                     print(f'"{input_file}" is a directory')
                 continue
 
+            # Was it already processed?
             path = os.path.split(input_file)
             ext = os.path.splitext(input_file)[-1]
             if not args.force and is_already_hash(path[-1], len(hasher.hexdigest())):
@@ -98,6 +108,7 @@ def process(args, algo):
                     print(f'Skipping "{input_file}"')
                 continue
 
+            # Calculate hash
             with open(input_file, 'rb') as f:
                 while True:
                     chunk = f.read(BUF_SIZE)
@@ -109,7 +120,8 @@ def process(args, algo):
             if os.path.isfile(out_file) and not args.force_rename:
                 print(f'"{out_file}" already exists, skipping "{input_file}"')
             else:
-                os.rename(input_file, out_file)
+                if not args.dry_run:
+                    os.rename(input_file, out_file)
                 print(f'"{input_file}" -> "{out_file}"')
         except OSError:
             sys.stderr.write(f'Cannot access to {input_file}\n')
